Separate code, few fixes, test adjustments
