+++ compiler/include/resolution.h
@@ -332,7 +332,4 @@ void stopGenerousResolutionForErrors();
 
 // Return the array element type, or NULL if not an array
 Type* arrayElementType(Type* arrayType);
-bool adjustAutoLocalAccessStatic(CallExpr *call, Immediate *imm);
-bool adjustAutoLocalAccessDynamic(CallExpr *call);
-
 #endif
