+++ compiler/passes/Makefile.share
@@ -21,6 +21,7 @@ GEN_HEADERS =                                              \
 
 PASSES_SRCS =                                              \
         addInitGuards.cpp                                  \
+        autoLocalAccess.cpp                                \
         buildDefaultFunctions.cpp                          \
         checkNormalized.cpp                                \
         checkParsed.cpp                                    \
