+++ test/optimizations/autoLocalAccess/dotDomDeclaration.good
@@ -4,6 +4,9 @@
 Iterated symbol
 	dotDomDeclaration.chpl:4
 
+Loop is suitable for further analysis
+	dotDomDeclaration.chpl:14
+
 Potential access
 	dotDomDeclaration.chpl:15
 
@@ -61,6 +64,9 @@ Iterated over .domain of
 where its domain is
 	dotDomDeclaration.chpl:21
 
+Loop is suitable for further analysis
+	dotDomDeclaration.chpl:31
+
 Potential access
 	dotDomDeclaration.chpl:32
 
@@ -100,6 +106,9 @@ Iterated over .domain of
 where its domain is
 	dotDomDeclaration.chpl:38
 
+Loop is suitable for further analysis
+	dotDomDeclaration.chpl:48
+
 Potential access
 	dotDomDeclaration.chpl:49
 
@@ -138,6 +147,9 @@ Iterated over .domain of
 
 where its domain is
 	dotDomDeclaration.chpl:55
+
+Loop is suitable for further analysis
+	dotDomDeclaration.chpl:65
 
 Potential access
 	dotDomDeclaration.chpl:66
