+++ test/optimizations/autoLocalAccess/interveningForallOrOn.good
@@ -10,6 +10,9 @@
 Iterated symbol
 	interveningForallOrOn.chpl:5
 
+Loop is suitable for further analysis
+	interveningForallOrOn.chpl:11
+
 **** End forall ****
 	interveningForallOrOn.chpl:11
 
@@ -18,6 +21,9 @@ Iterated symbol
 
 Iterated symbol
 	interveningForallOrOn.chpl:19
+
+Loop is suitable for further analysis
+	interveningForallOrOn.chpl:25
 
 **** End forall ****
 	interveningForallOrOn.chpl:25
