+++ test/optimizations/autoLocalAccess/regularDeclaration2D.good
@@ -4,6 +4,9 @@
 Iterated symbol
 	regularDeclaration2D.chpl:4
 
+Loop is suitable for further analysis
+	regularDeclaration2D.chpl:14
+
 Potential access
 	regularDeclaration2D.chpl:15
 
@@ -57,6 +60,9 @@ Access base's domain is the iterator
 
 Iterated symbol
 	regularDeclaration2D.chpl:21
+
+Loop is suitable for further analysis
+	regularDeclaration2D.chpl:31
 
 Potential access
 	regularDeclaration2D.chpl:32
