+++ test/optimizations/autoLocalAccess/staticSuccessDynamicFail.chpl
@@ -1,0 +1,10 @@
+use BlockDist;
+
+var A = newBlockArr({0..10}, int);
+var B = newBlockArr({1..9}, int);
+
+forall i in B.domain {
+  B[i] = A[i];
+}
+
+writeln(B);
